0021-merge-two-sorted-lists/src/test/java/com/kacpermartela/leetcode/MergeTwoSortedListsTest.java
package com.kacpermartela.leetcode;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;

import org.junit.jupiter.api.Test;

class MergeTwoSortedListsTest {

    @Test
    void example1() {
        var list1 = new ListNode(1, new ListNode(2, new ListNode(4)));
        var list2 = new ListNode(1, new ListNode(3, new ListNode(4)));
        var expected =
                new ListNode(1, new ListNode(1, new ListNode(2, new ListNode(3, new ListNode(4, new ListNode(4))))));
        var actual = MergeTwoSortedLists.mergeTwoLists(list1, list2);
        assertThat(actual).isEqualTo(expected);
    }
}
